Created header for common functors
