+++ core/inst_handlers/v/RvvIntegerInsts.cpp
@@ -1,6 +1,8 @@
 #include <limits>
 
 #include "core/inst_handlers/v/RvvIntegerInsts.hpp"
+#include "core/inst_handlers/CommonFunctors.hpp"
+#include "core/inst_handlers/m/RvmFunctors.hpp"
 #include "core/inst_handlers/inst_helpers.hpp"
 #include "core/PegasusState.hpp"
 #include "core/ActionGroup.hpp"
